Extract Wizard service class to the core
Allowing it to be reused across OpenProject and not
just inside the storages module.
